meta: migrate deep-exports and dialcect-module-configuration
