+++ test/unit/connection-manager.test.ts
@@ -1,20 +1,22 @@
 import type { Connection, Sequelize } from '@sequelize/core';
-import chai from 'chai';
+import { expect } from 'chai';
 import type { SinonStub } from 'sinon';
 import sinon from 'sinon';
 import { createSequelizeInstance } from '../support';
 
-const expect = chai.expect;
-
 describe('connection manager', () => {
+  let connection: Connection;
+  let sequelize: Sequelize;
+
+  beforeEach(() => {
+    connection = {};
+    sequelize = createSequelizeInstance();
+  });
+
   describe('_connect', () => {
-    let connection: Connection;
-    let sequelize: Sequelize;
     let connectStub: SinonStub;
 
     beforeEach(() => {
-      connection = {};
-      sequelize = createSequelizeInstance();
       connectStub = sinon.stub(sequelize.connectionManager, 'connect').resolves(connection);
     });
 
@@ -52,14 +54,9 @@ describe('connection manager', () => {
   });
 
   describe('_disconnect', () => {
-    let connection: Connection;
-    let sequelize: Sequelize;
     let disconnectStub: SinonStub;
 
     beforeEach(() => {
-      connection = {};
-      sequelize = createSequelizeInstance();
-
       disconnectStub = sinon.stub(sequelize.connectionManager, 'disconnect');
     });
